+++ appcheck/firebase-appcheck-debug/src/main/java/com/google/firebase/appcheck/debug/internal/DebugAppCheckProvider.java
@@ -19,20 +19,22 @@ import static com.google.android.gms.common.internal.Preconditions.checkNotNull;
 import android.annotation.SuppressLint;
 import android.util.Log;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.TaskCompletionSource;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.annotations.concurrent.Background;
+import com.google.firebase.annotations.concurrent.Blocking;
 import com.google.firebase.appcheck.AppCheckProvider;
 import com.google.firebase.appcheck.AppCheckToken;
+import com.google.firebase.appcheck.debug.InternalDebugSecretProvider;
 import com.google.firebase.appcheck.internal.DefaultAppCheckToken;
 import com.google.firebase.appcheck.internal.NetworkClient;
 import com.google.firebase.appcheck.internal.RetryManager;
+import com.google.firebase.inject.Provider;
 import java.util.UUID;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.concurrent.Executor;
 
 public class DebugAppCheckProvider implements AppCheckProvider {
 
@@ -40,20 +42,27 @@ public class DebugAppCheckProvider implements AppCheckProvider {
   private static final String UTF_8 = "UTF-8";
 
   private final NetworkClient networkClient;
-  private final ExecutorService backgroundExecutor;
+  private final Executor blockingExecutor;
   private final RetryManager retryManager;
   private final Task<String> debugSecretTask;
 
-  // TODO(b/258273630): Migrate to go/firebase-android-executors
-  @SuppressLint("ThreadPoolCreation")
-  public DebugAppCheckProvider(@NonNull FirebaseApp firebaseApp, @Nullable String debugSecret) {
+  public DebugAppCheckProvider(
+      @NonNull FirebaseApp firebaseApp,
+      @NonNull Provider<InternalDebugSecretProvider> debugSecretProvider,
+      @Background Executor backgroundExecutor,
+      @Blocking Executor blockingExecutor) {
     checkNotNull(firebaseApp);
     this.networkClient = new NetworkClient(firebaseApp);
-    this.backgroundExecutor = Executors.newCachedThreadPool();
+    this.blockingExecutor = blockingExecutor;
     this.retryManager = new RetryManager();
+
+    String debugSecret = null;
+    if (debugSecretProvider.get() != null) {
+      debugSecret = debugSecretProvider.get().getDebugSecret();
+    }
     this.debugSecretTask =
         debugSecret == null
-            ? determineDebugSecret(firebaseApp, this.backgroundExecutor)
+            ? determineDebugSecret(firebaseApp, backgroundExecutor)
             : Tasks.forResult(debugSecret);
   }
 
@@ -61,10 +70,10 @@ public class DebugAppCheckProvider implements AppCheckProvider {
   DebugAppCheckProvider(
       @NonNull String debugSecret,
       @NonNull NetworkClient networkClient,
-      @NonNull ExecutorService backgroundExecutor,
+      @NonNull Executor blockingExecutor,
       @NonNull RetryManager retryManager) {
     this.networkClient = networkClient;
-    this.backgroundExecutor = backgroundExecutor;
+    this.blockingExecutor = blockingExecutor;
     this.retryManager = retryManager;
     this.debugSecretTask = Tasks.forResult(debugSecret);
   }
@@ -72,7 +81,7 @@ public class DebugAppCheckProvider implements AppCheckProvider {
   @VisibleForTesting
   @NonNull
   static Task<String> determineDebugSecret(
-      @NonNull FirebaseApp firebaseApp, @NonNull ExecutorService executor) {
+      @NonNull FirebaseApp firebaseApp, @NonNull Executor executor) {
     TaskCompletionSource<String> taskCompletionSource = new TaskCompletionSource<>();
     executor.execute(
         () -> {
@@ -103,7 +112,7 @@ public class DebugAppCheckProvider implements AppCheckProvider {
             task -> {
               ExchangeDebugTokenRequest request = new ExchangeDebugTokenRequest(task.getResult());
               return Tasks.call(
-                  backgroundExecutor,
+                  blockingExecutor,
                   () ->
                       networkClient.exchangeAttestationForAppCheckToken(
                           request.toJsonString().getBytes(UTF_8),