+++ appcheck/firebase-appcheck-safetynet/src/test/java/com/google/firebase/appcheck/safetynet/internal/SafetyNetAppCheckProviderTest.java
@@ -37,8 +37,8 @@ import com.google.firebase.appcheck.internal.AppCheckTokenResponse;
 import com.google.firebase.appcheck.internal.DefaultAppCheckToken;
 import com.google.firebase.appcheck.internal.NetworkClient;
 import com.google.firebase.appcheck.internal.RetryManager;
+import com.google.firebase.concurrent.TestOnlyExecutors;
 import java.io.IOException;
-import java.util.concurrent.ExecutorService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -62,7 +62,6 @@ public class SafetyNetAppCheckProviderTest {
   private static final String TIME_TO_LIVE = "3600s";
 
   private FirebaseApp firebaseApp;
-  private ExecutorService backgroundExecutor = MoreExecutors.newDirectExecutorService();
   @Mock private GoogleApiAvailability mockGoogleApiAvailability;
   @Mock private SafetyNetClient mockSafetyNetClient;
   @Mock private NetworkClient mockNetworkClient;
@@ -84,7 +83,8 @@ public class SafetyNetAppCheckProviderTest {
     assertThrows(
         NullPointerException.class,
         () -> {
-          new SafetyNetAppCheckProvider(null);
+          new SafetyNetAppCheckProvider(
+              null, TestOnlyExecutors.background(), TestOnlyExecutors.blocking());
         });
   }
 
@@ -95,7 +95,11 @@ public class SafetyNetAppCheckProviderTest {
         .thenReturn(ConnectionResult.SERVICE_MISSING);
     SafetyNetAppCheckProvider provider =
         new SafetyNetAppCheckProvider(
-            firebaseApp, mockNetworkClient, mockGoogleApiAvailability, backgroundExecutor);
+            firebaseApp,
+            mockNetworkClient,
+            mockGoogleApiAvailability,
+            TestOnlyExecutors.background(),
+            TestOnlyExecutors.blocking());
     assertThat(provider.getSafetyNetClientTask().isSuccessful()).isFalse();
   }
 
@@ -105,7 +109,11 @@ public class SafetyNetAppCheckProviderTest {
         .thenReturn(ConnectionResult.SERVICE_MISSING);
     SafetyNetAppCheckProvider provider =
         new SafetyNetAppCheckProvider(
-            firebaseApp, mockNetworkClient, mockGoogleApiAvailability, backgroundExecutor);
+            firebaseApp,
+            mockNetworkClient,
+            mockGoogleApiAvailability,
+            TestOnlyExecutors.background(),
+            TestOnlyExecutors.blocking());
     assertThat(provider.getSafetyNetClientTask().isSuccessful()).isFalse();
 
     Task<AppCheckToken> tokenTask = provider.getToken();
@@ -120,7 +128,7 @@ public class SafetyNetAppCheckProviderTest {
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            TestOnlyExecutors.blocking(),
             mockRetryManager);
     assertThat(provider.getSafetyNetClientTask().getResult()).isEqualTo(mockSafetyNetClient);
 
@@ -142,7 +150,7 @@ public class SafetyNetAppCheckProviderTest {
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            TestOnlyExecutors.blocking(),
             mockRetryManager);
     assertThrows(
         NullPointerException.class,
@@ -160,7 +168,7 @@ public class SafetyNetAppCheckProviderTest {
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            TestOnlyExecutors.blocking(),
             mockRetryManager);
     assertThrows(
         IllegalArgumentException.class,
@@ -177,7 +185,7 @@ public class SafetyNetAppCheckProviderTest {
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            TestOnlyExecutors.blocking(),
             mockRetryManager);
     Task<AppCheckToken> task =
         provider.exchangeSafetyNetAttestationResponseForToken(mockSafetyNetAttestationResponse);
@@ -193,12 +201,13 @@ public class SafetyNetAppCheckProviderTest {
     when(mockAppCheckTokenResponse.getToken()).thenReturn(APP_CHECK_TOKEN);
     when(mockAppCheckTokenResponse.getTimeToLive()).thenReturn(TIME_TO_LIVE);
 
+    // TODO(b/258273630): Use TestOnlyExecutors instead of MoreExecutors.directExecutor().
     SafetyNetAppCheckProvider provider =
         new SafetyNetAppCheckProvider(
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            MoreExecutors.directExecutor(),
             mockRetryManager);
     Task<AppCheckToken> task =
         provider.exchangeSafetyNetAttestationResponseForToken(mockSafetyNetAttestationResponse);
@@ -219,12 +228,13 @@ public class SafetyNetAppCheckProviderTest {
             any(), eq(NetworkClient.SAFETY_NET), eq(mockRetryManager)))
         .thenThrow(new IOException());
 
+    // TODO(b/258273630): Use TestOnlyExecutors instead of MoreExecutors.directExecutor().
     SafetyNetAppCheckProvider provider =
         new SafetyNetAppCheckProvider(
             firebaseApp,
             mockSafetyNetClient,
             mockNetworkClient,
-            backgroundExecutor,
+            MoreExecutors.directExecutor(),
             mockRetryManager);
     Task<AppCheckToken> task =
         provider.exchangeSafetyNetAttestationResponseForToken(mockSafetyNetAttestationResponse);